requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
colorama>=0.4.4
pyyaml>=5.4.1
tldextract>=3.1.0
validators>=0.18.2
urllib3>=1.26.5 